refactoring: checking by checklist

+++ controllers/movie.js
@@ -1,5 +1,7 @@
 const Movie = require('../models/movie');
 const BadRequestError = require('../errors/BadRequestError');
+const NotFoundError = require('../errors/NotFoundError');
+const NotRootError = require('../errors/NotRootError');
 
 const getMovies = (req, res, next) => {
   const owner = req.user._id;
@@ -13,8 +15,17 @@ const getMovies = (req, res, next) => {
 
 const createMovie = (req, res, next) => {
   const {
-    country, director, duration, year, description,
-    image, trailerLink, nameRU, nameEN, thumbnail, movieId,
+    country,
+    director,
+    duration,
+    year,
+    description,
+    image,
+    trailerLink,
+    nameRU,
+    nameEN,
+    thumbnail,
+    movieId,
   } = req.body;
   const owner = req.user._id;
 
@@ -58,11 +69,17 @@ const createMovie = (req, res, next) => {
 };
 
 const deleteMovie = (req, res, next) => {
-  Movie.findById(req.params.movieId).select('+owner')
-    .then(() => {
-      Movie.findByIdAndRemove(req.params.movieId).select('-owner')
-        .then((movieIsDeleted) => res.send(movieIsDeleted))
-        .catch(next);
+  Movie.findById(req.params.movieId).select('+owner').orFail(new NotFoundError('Фильм с таким данными не найден'))
+    .then((movie) => {
+      const user = String(req.user._id);
+      const movieOwner = String(movie.owner);
+      if (user === movieOwner) {
+        Movie.findByIdAndRemove(req.params.movieId).select('-owner')
+          .then((movieIsDeleted) => res.send(movieIsDeleted))
+          .catch(next);
+      } else {
+        next(new NotRootError('Вы можете удалять только собственные фильмы'));
+      }
     })
     .catch((error) => {
       if (error.name === 'CastError') {
